azure-services-vision/requirements.txt
azure-ai-vision-imageanalysis
python-env
dotenv